pyproject.toml
[project]
name = "flask-webserver"
version = "0.1.0"
description = ""
authors = [
    {name = "a.cervelione",email = "a.cervelione@reply.it"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pyspark (>=3.5.5,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
]

[tool.poetry.group.test.dependencies]
fastapi = "^0.115.11"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
requests = "^2.32.3"
pyspark = "^3.5.5"
pyjwt = "^2.10.1"
pandas = "^2.2.3"
pytest = "^7.2.0"
pytest-azurepipelines = "^1.0.4"
pytest-cov = "^4.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
